+++ src/Moip.php
@@ -4,13 +4,17 @@ namespace Moip;
 
 use Moip\Contracts\Authentication;
 use Moip\Resource\Account;
+use Moip\Resource\BankAccount;
 use Moip\Resource\Customer;
 use Moip\Resource\Entry;
+use Moip\Resource\Keys;
 use Moip\Resource\Multiorders;
 use Moip\Resource\NotificationPreferences;
 use Moip\Resource\Orders;
 use Moip\Resource\Payment;
+use Moip\Resource\Refund;
 use Moip\Resource\Transfers;
+use Moip\Resource\WebhookList;
 use Requests_Session;
 
 /**
@@ -44,7 +48,7 @@ class Moip
      *
      * @const string
      */
-    const CLIENT_VERSION = '1.3.1';
+    const CLIENT_VERSION = '2.2.0';
 
     /**
      * Authentication that will be added to the header of request.
@@ -200,6 +204,46 @@ class Moip
     public function notifications()
     {
         return new NotificationPreferences($this);
+    }
+
+    /**
+     * Create a new WebhookList instance.
+     *
+     * @return WebhookList
+     */
+    public function webhooks()
+    {
+        return new WebhookList($this);
+    }
+
+    /**
+     * Create a new Keys instance.
+     *
+     * @return Keys
+     */
+    public function keys()
+    {
+        return new Keys($this);
+    }
+
+    /**
+     * Create a new Refund instance.
+     *
+     * @return Refund
+     */
+    public function refunds()
+    {
+        return new Refund($this);
+    }
+
+    /**
+     * Create a new BankAccount instance.
+     *
+     * @return BankAccount
+     */
+    public function bankaccount()
+    {
+        return new BankAccount($this);
     }
 
     /**