+++ tests/Resource/CustomerTest.php
@@ -47,6 +47,8 @@ class CustomerTest extends TestCase
         $this->assertEquals($customer_original->getFullname(), $customer->getFullname());
         $this->assertEquals($customer_original->getPhoneNumber(), $customer->getPhoneNumber());
         $this->assertEquals($customer_original->getBirthDate(), $customer->getBirthDate());
+        $this->assertEquals($customer_original->getTaxDocumentType(), $customer->getTaxDocumentType());
+        $this->assertEquals($customer_original->getTaxDocumentNumber(), $customer->getTaxDocumentNumber());
     }
 
     /**
@@ -65,5 +67,31 @@ class CustomerTest extends TestCase
         $this->assertEquals($customer_original->getShippingAddress()->state, $customer->getShippingAddress()->state);
         $this->assertEquals($customer_original->getShippingAddress()->country, $customer->getShippingAddress()->country);
         $this->assertEquals($customer_original->getShippingAddress()->zipCode, $customer->getShippingAddress()->zipCode);
+    }
+
+    /**
+     * MoipTest create credit card for customer.
+     */
+    public function testCreateCreditCard()
+    {
+        $this->mockHttpSession($this->body_client);
+        $customer = $this->createCustomer()->create();
+
+        $this->mockHttpSession($this->body_add_credit_card);
+
+        $creditCard = $this->moip->customers()->creditCard()
+            ->setExpirationMonth('05')
+            ->setExpirationYear(2018)
+            ->setNumber('4012001037141112')
+            ->setCVC('123')
+            ->setFullName('Jose Portador da Silva')
+            ->setBirthDate('1988-12-30')
+            ->setTaxDocument('CPF', '33333333333')
+            ->setPhone('55', '11', '66778899')
+            ->create($customer->getId());
+
+        $this->assertNotEmpty($creditCard->getId());
+        $this->assertEquals($creditCard->getFirst6(), '401200');
+        $this->assertEquals($creditCard->getLast4(), '1112');
     }
 }